requirements.txt
aioconsole==0.4.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
chardet==4.0.0
discord.py==1.7.3
flake8==4.0.1
idna==3.3
mccabe==0.6.1
multidict==6.0.2
mypy==0.942
mypy-extensions==0.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
yarl==1.7.2
